requirements.txt
streamlit==0.69.0
Pillow
opencv-python==4.4.0.44
opencv-contrib-python
opencv-contrib-python-headless
scikit-image==0.17.2
imutils==0.5.3


